requirements.txt
pyodbc
fastapi
pydantic
sqlalchemy
passlib
bcrypt
uvicorn
python-jose
pyjwt
python-multipart